requirements.txt
customtkinter==5.2.2
pymupdf==1.24.10
pillow==10.4.0
tkinterdnd2==0.4.2
